chore: fix issues after merge, add comments
