+++ Sources/RInAppMessaging/TooltipDispatcher.swift
@@ -61,7 +61,7 @@ internal class TooltipDispatcher: TooltipDispatcherType, ViewListenerObserver {
         guard let resImgUrlString = tooltip.tooltipData?.imageUrl,
               let resImgUrl = URL(string: resImgUrlString)
         else {
-            // TOOLTIP: display premission?
+            // TOOLTIP: display permission?
             return
         }
